+++ src/main/java/com/johnsunday/app/controller/IPayrollController.java
@@ -8,16 +8,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 public interface IPayrollController<Payroll> {
 
-	public ResponseEntity<?> getAllPayrollByEmployeeId(@PathVariable Integer employeeId);
+	public ResponseEntity<?> getAllPayrollByEmployeeId(@PathVariable Long employeeId);
 
 	public ResponseEntity<?> getAllPayroll();
 
-	public ResponseEntity<?> getPayrollById(@PathVariable Integer payrollId);
+	public ResponseEntity<?> getPayrollById(@PathVariable Long payrollId);
 
 	public ResponseEntity<?> savePayroll(@RequestBody @Valid Payroll payroll);
 
 	public ResponseEntity<?> updatePayroll(@RequestBody @Valid Payroll payroll,
-			@PathVariable Integer payrollId);
+			@PathVariable Long payrollId);
 
-	public ResponseEntity<?> deletePayroll(@PathVariable Integer payrollId);
+	public ResponseEntity<?> deletePayroll(@PathVariable Long payrollId);
 }
