+++ src/main/java/com/johnsunday/app/dao/IEmployeeDao.java
@@ -8,9 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.johnsunday.app.entity.Employee;
 
 @Repository
-public interface IEmployeeDao extends JpaRepository<Employee,Integer> {
+public interface IEmployeeDao extends JpaRepository<Employee, Long> {
 
-	//@Query(value="SELECT * FROM employee WHERE employee.name=?1 and employee.surname=?2", nativeQuery=true)
-	Optional<Employee> findByNameAndSurnameAllIgnoreCase(String name,String surname);
+	// @Query(value="SELECT * FROM employee WHERE employee.name=?1 and
+	// employee.surname=?2", nativeQuery=true)
+	Optional<Employee> findByNameAndSurnameAllIgnoreCase(String name, String surname);
+
 	Optional<Employee> findByEmail(String email);
 }
