+++ src/main/java/com/johnsunday/app/security/controller/IUserController.java
@@ -11,12 +11,12 @@ public interface IUserController {
 
 	public ResponseEntity<?> getAllUser();
 
-	public ResponseEntity<?> getUserById(@PathVariable Integer userId);
+	public ResponseEntity<?> getUserById(@PathVariable Long userId);
 
 	public ResponseEntity<?> saveUser(@RequestBody @Valid ExpenseUser user);
 
 	public ResponseEntity<?> updateUser(@RequestBody @Valid ExpenseUser user,
-			@PathVariable Integer userId);
+			@PathVariable Long userId);
 
-	public ResponseEntity<?> deleteUser(@PathVariable Integer userId);
+	public ResponseEntity<?> deleteUser(@PathVariable Long userId);
 }
